backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
motor==3.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
celery==5.3.6
redis==5.0.4
crewai==0.30.0
crewai-tools==0.1.7
langchain-google-genai==1.0.2
pypdf==4.2.0
langchain-community==0.0.38
markdown==3.6
pymongo==4.6.1
python-multipart==0.0.9